requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
curl_cffi==0.6.0b9
dataclasses-json==0.6.3
distro==1.9.0
duckduckgo_search==4.4.3
ecdsa==0.18.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.110.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.4
langchain-community==0.0.19
langchain-core==0.1.22
langchain-openai==0.0.5
langsmith==0.0.87
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.3
openai==1.10.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pinecone-client==3.0.2
protobuf==4.25.3
pyarrow==15.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
Pygments==2.17.2
pyodbc==5.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
streamlit==1.31.1
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.0
uvicorn==0.27.1
validators==0.22.0
watchdog==4.0.0
yarl==1.9.4
zipp==3.17.0
